# feat(cli): isCIEnv flag