--- a/packages/cli/src/command-config/default-flag-args.ts
+++ b/packages/cli/src/command-config/default-flag-args.ts
@@ -40,6 +40,7 @@ export const defaultFlagArgs: ITBConfig = {
   emulateDeviceOrientation: 'vertical',
   dest: '',
   regressionThreshold: 100,
+  isCIEnv: false,
 };
 
 // specify with --headless flag