--- a/packages/cli/src/command-config/tb-config.ts
+++ b/packages/cli/src/command-config/tb-config.ts
@@ -41,6 +41,7 @@ export interface ITBConfig {
   [EXPERIMENT_ENV_OVERRIDE_ATTR]?: IBenchmarkEnvironmentOverride;
   [key: string]: any;
   cookie?: string;
+  isCIEnv?: boolean | string;
 }
 
 export interface IHARServer {