- This boolean flag `isCIEnv` for the `compare` cmd will allow for CI Environment stdout reports to be more streamlined and removes all tables.

Resulting in:

```console
SUCCESS!     20 test samples were taken.

!! ALERT      Regression found exceeding the set regression threshold of 100ms

Benchmark Results Summary    

Red color means there was a regression.
Green color means there was an improvement.

duration phase estimated difference +1007ms [1001ms to 1014ms]
jquery phase estimated difference +2ms [0ms to 4ms]
ember phase no difference
application phase estimated difference +1004ms [1003ms to 1004ms]
routing phase no difference
transition phase no difference
render phase no difference
afterRender phase estimated difference -3ms [-4ms to 0ms]
```